custom_chat_gpt/requirements.txt
python-dotenv
openai
pandas
numpy
pyyaml
flask
markdown